refactored packages for clarity
