--- src/main/Java/data_access/Tempfile.java
@@ -1,4 +1,0 @@
-package data_access;
-
-public class Tempfile {
-}
