+++ src/main/Java/data_access/geo/GeocodingService.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.geo;
 
 public interface GeocodingService {
     Location getCoordinates(String query);
