+++ src/main/Java/data_access/geo/Location.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.geo;
 
 public class Location {
     private final double latitude;
